# Use id-arena to store `Expr` and `Type`