--- a/mimium-lang/src/types.rs
+++ b/mimium-lang/src/types.rs
@@ -1,5 +1,7 @@
 use std::fmt;
 
+use id_arena::Id;
+
 use crate::{ast::Symbol, format_vec};
 
 use super::utils::miniprint::MiniPrint;