Address  #12

## Overview

This pull request introduces the following storage into `SessionGlobals`

1. `expr_storage` to store `Expr`
2. `type_storage` to store `Type`
3. `span_storage` to store `Span` with the key of the ID of `Expr` or `Type`

As a result,

* Since `Expr` and `Type` are immediately converted into a copyable Id, so `Box` is no longer needed for recursive definitions
* Since all `Span`s are recorded in `span_storage`, `WithMeta` is no longer needed

## Remaining questions

### ID for `Span` of `TypePattern` and `TypedId`

Since `TypePattern` and `TypedId` has only one `Type`, I decided to use this ID for the key of the corresponding `Span`. However, of course, this is not accurate. For example, a `TypedId` represents a `Symbol` and a colon in addition to the `Type`. I guess this isn't a big problem at the moment, but the error message can be confusing because of this.

``` rust
    let x:float = 1.0
//        ^^^^^ 1
//      ^^^^^^^ 2
```

Maybe I could add something like `type_pattern_storage` and `typed_id_storage`, and use the key. But, I'd like to keep this pull request simple because this is already a bit complex.