@yutannihilation 2024-08-21
Currently, the test fails with the following error. This is weird because, if I investigate manually, there's no difference between the actual `Expr`s of these two `ExprId`s.

```
---- compiler::parser::test::test_tuple stdout ----
thread 'compiler::parser::test::test_tuple' panicked at mimium-lang\src\compiler\parser\test.rs:266:5:
assertion `left == right` failed
  left: WithMeta(Tuple([ExprId(Id { idx: 3 }), ExprId(Id { idx: 4 })]), 0..10)
 right: WithMeta(Tuple([ExprId(Id { idx: 0 }), ExprId(Id { idx: 1 })]), 0..10)
```