@tomoyanonymous 2024-08-21
Thanks. This really needed to address the type inference issue...
Having span as `BTreeMap` with `ExprId` is what I was expected for the implementation.

> Currently, the test fails with the following error. This is weird because, if I investigate manually, there's no difference between the actual Exprs of these two ExprIds.

`ExprId` should not derive `PartialEq` and `Eq` but should impl PartialEq manually by comparing actual `Expr` in the comparison.
There maybe the situation that we need to compare internal id but we can still do it by accessing internal `id_arena::Id<Expr>` directly.