@tomoyanonymous 2024-08-21
`Pattern` and `Type` should be allocated on the session_global similarly...