@yutannihilation 2024-08-21
> `ExprId` should not derive `PartialEq` and `Eq` but should impl PartialEq manually by comparing actual `Expr` in the comparison.

Oh, thanks for your insight!!! I'll confirm this tomorrow.

> `Pattern` and `Type` should be allocated on the session_global similarly...

I agree. I'm not immediately sure if these need dedicated storage for each, or a variant of `Expr` enum  (e.g. `Expr::TypePattern(TypePattern)`), but these probably need to be stored in session_globals somehow.