@tomoyanonymous 2024-08-21
>I'm not immediately sure if these need dedicated storage for each, or a variant of Expr enum (e.g. Expr::TypePattern(TypePattern)

I more prefer to leaving `Pattern` as an individual recursive enum, Although we need to manage different storages, it ensures `Let` must hold `TypedPattern` in its left-hand, so it reduces the cost of validation.