@yutannihilation 2024-08-21
If I implement `PartialEq` manually, it works fine. Thanks for the advice! (What I said "weird" was simply because I assumed wrongly that id-arena removes duplication automatically...)