@tomoyanonymous 2024-08-22
I realized that the rust docs says
`The methods of this trait must be consistent with each other and with those of [PartialEq](https://doc.rust-lang.org/beta/std/cmp/trait.PartialEq.html).`

https://doc.rust-lang.org/beta/std/cmp/trait.PartialOrd.html

Now the key for `BtreeMap` is `NodeId`, not `ExprNodeId
` itself, so `ExprNodeId` should not derive `PartialOrd` and `Ord` too.