@yutannihilation 2024-08-22
I chose the direction to treat `Type` and `Expr` as the same. While this works, I'm getting to feel I was wrong. 

- `Type` can be used repeatedly. If one type and another have a same type, they should have the same ID rather than issuing a new one by `Arena::alloc()` every time.
- `Type` doesn't always have `Span`.

Probably I need to go back and forth several times.