@tomoyanonymous 2024-08-22
> `Type` doesn't always have `Span`.

I think the BTreeMap from NodeId to Span can be extended to any kind of metadata. 

For example, I'm going to refactor the type inference code so that MIR generator later can refer to the map from `ExprNodeId` to `Type` after the entire inference finished.

Duplicating same values with different ids will not be an immidiate problem. 

Maybe, It is more clear that the Type values issued during the inference should be distinct from those of TypedPattern (Type inference context have an individual Arena) but it is ok to be allocated on the same Arena.