@yutannihilation 2024-08-22
> Maybe, It is more clear that the Type values issued during the inference should be distinct from those of TypedPattern (Type inference context have an individual Arena) but it is ok to be allocated on the same Arena.

Thanks for sharing the plan. This sounds good to me! I agree this duplication isn't any problem (except for some possible performance overhead).

I think I don't understand the whole picture here, but probably I can finish this pull request today or tomorrow.