@yutannihilation 2024-08-23
This should be basically ready for review now. I updated the main description.

One problem I'm seeing is that this pull request makes the benchmark tests fail with a segmentation fault. It seems this happens only on benchmark, so I don't think this is a blocker. If I run the same Mimium code by `cargo test` or by `cargo run`, nothing happens. I'll keep investigating.

```
warning: `mimium-lang` (lib) generated 40 warnings (run `cargo fix --lib -p mimium-lang` to apply 7 suggestions)
    Finished `bench` profile [optimized] target(s) in 0.04s
     Running benches/bench.rs (target/release/deps/bench-d16f0f1ff08fbac1)

running 1 test
error: bench failed, to rerun pass `-p mimium-lang --bench bench`

Caused by:
  process didn't exit successfully: `/home/yutani/repo/mimium-rs/target/release/deps/bench-d16f0f1ff08fbac1 'tests::parse::bench_many_symbols3' --exact --show-output --bench` (signal: 11, SIGSEGV: invalid memory reference)

 *  The terminal process "cargo 'bench', '--package', 'mimium-lang', '--bench', 'bench', '--', 'tests::parse::bench_many_symbols3', '--exact', '--show-output'" failed to launch (exit code: 101). 
 *  Terminal will be reused by tasks, press any key to close it. 
```

```
fn f0(x3:float) {
  let x3_0 = x3
  let x3_1 = x3
  let x3_2 = x3
  (((x3 * 0.1 + x3_0) * 0.1 + x3_1) * 0.1 + x3_2)
}



fn f1(x3:float) {
  let x3_0 = x3
  let x3_1 = x3
  let x3_2 = x3
  (((x3 * 0.1 + x3_0) * 0.1 + x3_1) * 0.1 + x3_2)
}



fn f2(x3:float) {
  let x3_0 = x3
  let x3_1 = x3
  let x3_2 = x3
  (((x3 * 0.1 + x3_0) * 0.1 + x3_1) * 0.1 + x3_2)
}


fn dsp() {
  f0(1.0) + f1(1.0) + f2(1.0)
}
```