@tomoyanonymous 2024-08-23
> One problem I'm seeing is that this pull request makes the benchmark tests fail with a segmentation fault. It seems this happens only on benchmark, so I don't think this is a blocker. If I run the same Mimium code by cargo test or by cargo run, nothing happens. I'll keep investigating.

Did you tried with `--profile bench` option?