@yutannihilation 2024-08-23
I didn't. Now I tried it, and it runs fine without any errors.

However, let me correct a bit. This code can reproduce the segfault.

- If I reduce the repetition of `emit_mir()`, it doesn't fail
- If I increase the amount of `content`, it fails with fewer repetition.

So, this is probably a real issue I need to address before merging. Probably, this might take some time. Please feel free to merge #28 first. I can retry after that...

``` rust
        #[test]
        fn t() {
            let content = "

...snip... (the code above)

";
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
            emit_mir(content).expect("ok");
        }
```

If I execute the above code with VS Code's Debug run, it points this line of `mimium-lang/src/compiler/mirgen/recursecheck.rs`. I have no idea in what situation this can happen.

![image](https://github.com/user-attachments/assets/e8b8beb9-f67f-495a-abcc-3c87a45eb519)