@tomoyanonymous 2024-08-23
Hmm, that test won't fail on macOS how many times I duplicate`emit_mir`. My fix in the review above does not change the behaviour too.

I guess it happens on `std::transmute<&Expr,&Expr>` in `to_expr` for extending lifetime? 

This transmute is done on [`rustc_span`](https://github.com/rust-lang/rust/blob/b5723af3457b9cd3795eeb97e9af2d34964854f2/compiler/rustc_span/src/symbol.rs#L2324) too but its content type is `&str`, not a recursive type. 
It causes some difference maybe?