@yutannihilation 2024-08-23
Okay, I give up avoiding the cloning costs. Now the benchmark doesn't fail.

Benchmark results:

**Before**
```
test tests::parse::bench_many_symbols10 ... bench:  22,420,550.00 ns/iter (+/- 2,571,804.00)
test tests::parse::bench_many_symbols3  ... bench:   2,301,567.50 ns/iter (+/- 56,266.75)
test tests::parse::bench_many_symbols5  ... bench:   5,573,355.00 ns/iter (+/- 183,844.00)
test tests::runtime::bench_multiosc10   ... bench:       3,429.20 ns/iter (+/- 109.64)
test tests::runtime::bench_multiosc15   ... bench:       5,100.33 ns/iter (+/- 125.19)
test tests::runtime::bench_multiosc5    ... bench:       1,748.98 ns/iter (+/- 48.91)
test tests::runtime::bench_multiosc7    ... bench:       2,396.01 ns/iter (+/- 54.89)
```


**After**
```
test tests::parse::bench_many_symbols10 ... bench:  16,587,370.00 ns/iter (+/- 2,167,060.00)
test tests::parse::bench_many_symbols3  ... bench:   1,954,627.50 ns/iter (+/- 68,536.00)
test tests::parse::bench_many_symbols5  ... bench:   4,715,255.00 ns/iter (+/- 472,429.50)
test tests::runtime::bench_multiosc10   ... bench:       3,365.50 ns/iter (+/- 109.28)
test tests::runtime::bench_multiosc15   ... bench:       4,992.07 ns/iter (+/- 113.04)
test tests::runtime::bench_multiosc5    ... bench:       1,681.22 ns/iter (+/- 44.98)
test tests::runtime::bench_multiosc7    ... bench:       2,359.02 ns/iter (+/- 39.85)
```